refactor: split Side to other file
