+++ src/components/molecules/Visualizer/Engine/Cesium/Box/index.tsx
@@ -1,20 +1,11 @@
-import {
-  Axis,
-  CallbackProperty,
-  Cartesian2,
-  Cartesian3,
-  Color,
-  Matrix4,
-  Plane as CesiumPlane,
-  PositionProperty,
-  TranslationRotationScale,
-} from "cesium";
-import React, { useMemo, FC, useEffect, memo, useState } from "react";
-import { Entity, PlaneGraphics } from "resium";
+import { Cartesian3, Color, Plane as CesiumPlane, TranslationRotationScale } from "cesium";
+import React, { useMemo, useEffect, memo, useState } from "react";
 
 import { LatLng, toColor } from "@reearth/util/value";
 
 import type { Props as PrimitiveProps } from "../../../Primitive";
+
+import { Side } from "./Side";
 
 export type Props = PrimitiveProps<Property>;
 
@@ -38,63 +29,6 @@ export type Property = {
     };
   } & BoxStyle<string>;
 };
-
-// ref: https://github.com/TerriaJS/terriajs/blob/cad62a45cbee98c7561625458bec3a48510f6cbc/lib/Models/BoxDrawing.ts#L1446-L1461
-function setPlaneDimensions(
-  boxDimensions: Cartesian3,
-  planeNormalAxis: Axis,
-  planeDimensions: Cartesian2,
-) {
-  if (planeNormalAxis === Axis.X) {
-    planeDimensions.x = boxDimensions.y;
-    planeDimensions.y = boxDimensions.z;
-  } else if (planeNormalAxis === Axis.Y) {
-    planeDimensions.x = boxDimensions.x;
-    planeDimensions.y = boxDimensions.z;
-  } else if (planeNormalAxis === Axis.Z) {
-    planeDimensions.x = boxDimensions.x;
-    planeDimensions.y = boxDimensions.y;
-  }
-}
-
-const Side: FC<{
-  id: string;
-  planeLocal: CesiumPlane;
-  trs: TranslationRotationScale;
-  style: BoxStyle<Color>;
-}> = memo(function SidePresenter({ id, planeLocal, style, trs }) {
-  const normalAxis = planeLocal.normal.x ? Axis.X : planeLocal.normal.y ? Axis.Y : Axis.Z;
-  const cbRef = useMemo(
-    () => new CallbackProperty(() => trs.translation, false) as unknown as PositionProperty,
-    [trs],
-  );
-  const [plane, dimension] = useMemo(() => {
-    const dimension = new Cartesian3();
-    setPlaneDimensions(trs.scale, normalAxis, dimension);
-    const scratchScaleMatrix = new Matrix4();
-    const scaleMatrix = Matrix4.fromScale(trs.scale, scratchScaleMatrix);
-    const plane = CesiumPlane.transform(
-      planeLocal,
-      scaleMatrix,
-      new CesiumPlane(Cartesian3.UNIT_Z, 0),
-    );
-    return [plane, dimension];
-  }, [trs, normalAxis, planeLocal]);
-
-  return (
-    <Entity id={id} position={cbRef}>
-      <PlaneGraphics
-        plane={plane}
-        dimensions={dimension}
-        fill={style.fill}
-        outline={style.outline}
-        material={style.fillColor}
-        outlineColor={style.outlineColor}
-        outlineWidth={style.outlineWidth}
-      />
-    </Entity>
-  );
-});
 
 // The 6 box sides defined as planes in local coordinate space.
 // ref: https://github.com/TerriaJS/terriajs/blob/cad62a45cbee98c7561625458bec3a48510f6cbc/lib/Models/BoxDrawing.ts#L161-L169
